♻️ [REFACTOR] 프로필 및 404 페이지 구조 개선
- ProfilePage 및 NotFound 페이지를 새로운 구조로 리팩토링하여 코드 가독성 향상
- ProfilePage를 profile 폴더로 이동하고, 관련 컴포넌트로 분리
- NotFound 페이지를 not-found 폴더로 이동하여 구조 정리
- 기존 ProfilePage 및 NotFound.tsx 파일 삭제
